iplot performance improvement by 4x. Yikes.
Every _repr_* method that exists is called on instantiation of a
IPython.core.display.HTML subclass. For Plotly,
_repr_png, _repr_svg, _repr_jpeg require a call to the image server
to generate the image. Combined, the 4 requests can make iplot commands
take 4 times as long as usual.

We mistakenly thought that the _repr_ commands were only called on demand,
e.g. on a call to `nbconvert`.

Unfortunately, this removes our support for `nbconvert`. An upcoming PR with
cached images should bring this support back.