Compile the `CLASS_NAMES_TO_OBJECT_NAMES` dict.
This is used to map a name like `Line` to the object `line` that exists
in graph reference.

For completeness, we then use `line` to lookup all the paths of `line`
in `OBJECTS` which tells us *all* the possible parents `line` can have
and all the attributes `line` can have given its parent.