Add public method to `get_object_info`.
The object is something like `marker` which *should* be looked up by
a path, E.g. `(‘traces’, ‘scatter’, ‘marker’)`. However, there are
caveats so we need to allow both a `path` and an `object_name` param.

`figure`, `data`, `layout`, and all the traces are special in that they
are either not strictly defined in the plot schema or they require
special functionality to gather their attributes.

`annotation` and `shape` are special in that their parent objects,
`annotations` and `shapes` both point to the same location in graph
reference. This means that we need *both* the name `annotation` and the
path to get our info.