Replace old `value_is_data` with `get_role`.
`value_is_data` used the old `INFO` dict and the value of an attribute
to decide whether something had the `data` role or not. However, we can
just replace this with a general `get_role` function that can give the
default role or the ‘implicit’ role based on the value.