Add in support for the `copy` module. (more)
This looks sketchier than it is :) Python allows you to define
`__copy__` and `__deepcopy__` methods which it will use when a user does
`copy.copy(obj)` or `copy.deepcopy(obj)`. We actually test for this
since we use the `copy` module internally, therefore this is a necessary
addition to keep tests passing.

See https://github.com/plotly/python-api/issues/291.