Add `memoize` decorator to improve performance.
The bottleneck for the `make_subplots` function was due to excessive
lookups in the `plot_schema`. These lookups are actually pretty
computation-intensive so caching these computations can give us a large
performance boost.

Note that Python 3.2+ has a new [`functools.lru_cache`](https://docs.python.org/3/library/functools.html#functools.lru_cache) which
can be used for this. HOWEVER, we support Python 2.7+ and I didn’t see a backport for it.

There are numerous `memoize` packages on PyPI, so many that I didn’t
want to commit to one. It’s fairly simple to write this and then we
don’t need another dependency.