Enable heartbeat in Stream.write
The (plot.ly streaming documentation)[https://plot.ly/streaming/] under the
Disconnections section states that:

> A connection can be maintained by writing a heartbeat within the 60 second
> window, a heartbeat is simply a newline

If an empty string (and maybe an empty dictionary too, untested) is passed to
Stream's `write` method it results in an empty row being created in the plot's
data rather than keeping the connection awake.  This commit aims to support
using the write method for sending a heartbeat.