Add missing rugplot property passing inside violin_no_colorscale
When trying to plot my violin, I noticed rugplot=False was not being respected. This is a proposed fix.
As a point of discussion, should the rugplot=True|False property name be normalized between the figure factories? For example, in _distplot.py the corresponding property name is "show_rug" instead of "rugplot".