CircleCI 2.0 migration (#1109)
Port of test suite to 2.0, with improvements

There is now a workflow for each Python version / test suite combination.  This way all of the jobs can run in parallel. I also refactored the `core` tests to pull out the tests that rely on `plot.ly` interaction. These are currently the slowest tests, and pulling them out means that the core tests run really fast. It also means that if there is an intermittent failure due to plot.ly communication, we can restart only the plot.ly tests for the particular version of Python that had the failure.

Updates:
 - [x] Split plot.ly tests into separate directory
 - [x] Add core and core+optional workflows for 2.7, 3.4, 3.5, 3.6, and 3.7 (These run pretty fast now without the plot.ly tests)
 - [x] plot.ly tests for 2.7, 3.4, and 3.7?
 - [x] Add pytest validator tests to optional (these tests need the extra dependencies)