Support for embedding plotlywidget instances that use numpy arrays (#1117)
* Null out _js2py properties after touch() so they don't end up in the saved model

* Added TypedArray js serialization to support saving embedded widget state.

For some reason, this required renaming 'buffer' in the
array representation object. Perhaps a conflict with naming conventions
in ipywidgets.  So the buffers are now stored in property named `value`

*  Several dynamic resizing improvements

 - Render initial empty figure on 'before-attach' event. This
   keeps the figure view from collapsing when opening a classic
   notebook that was saved with embedded widget state
 - Autoresize width even if height has been explicitly set
 - Autoresize in the classic notebook by responding to window
   resize events.

* Revert change to Python serializer.

If the new plotlywidget can work with current version of plotly.py then
we can put out a patch release of the widget only.

* Have plotlywidget accept multiple typed array representations

The buffer is named `buffer` when sent from plotly.py<=3.1.1,
but it is named `value` when restoring from saved widget state and when
receiving updates from plotly.py>=3.2.

* Emit ''plotlywidget-after-render' event when the widget has
finished rendering.

* Updated version to 0.2.2-rc.1 and published to npm

* Convert dispatch 'plotlywidget-after-render' to CustomEvent
dispatched on the document

* Version 0.2.2-rc.2 published to npm.