[WIP] Use circleci to build conda packages (#1154)
Use circleci to build conda packages

Previously the conda package was a noarch Python package that was not specific to Python version of operating system. This works fine, except for one problem.

Conda always ensures that all *.py files are byte-compiled to *.pyc when a package is installed, and this can take a long time (~5-10 minutes) for plotly.py since we have so many python files thanks to code generation.

To avoid this problem, this PR uses circleci to build conda packages for each Python version, and then converts that build to each operating system using conda convert. This way, each package includes the pre-compiled byte code for the correct version of Python. This results in a much fast installation experience (compared to conda noarch) and a faster initial import (compared to pip wheel).