Add skip_invalid kwarg to graph_objs constructors.
When set to True (Default is False) all invalid properties (invalid
by name or by value) will be skipped an no exception will be raised
due to schema violations.  The option applies recursively to all
child objects, but it only applies within the scope of the constructor.

Still TODO: add the skip_invalid option to the figure constructor.
propagate this state to trace, layout, and frames validators during
figure construction.