Introduce skip_invalid argument to figure (#1162)
* Add skip_invalid kwarg to graph_objs constructors.

When set to True (Default is False) all invalid properties (invalid
by name or by value) will be skipped an no exception will be raised
due to schema violations.  The option applies recursively to all
child objects, but it only applies within the scope of the constructor.

* Add skip_invalid kwarg to Figure and FigureWidget.

Added docstring documentatin of skip_invalid, and added
new tests.

When set to True (Default is False) all invalid properties (invalid
by name or by value) will be skipped an no exception will be raised
due to schema violations.  The option applies recursively to all
child objects, but it only applies within the scope of the constructor.