Added figure to_ordered_dict method (#1167)
Adds a new BaseFigure method, to_ordered_dict. This builds and returns a representation of the figure as a nested structure of OrderedDict and list instances. The OrderedDict keys are sorted alphabetically. This makes it possible for library users to iterate over the nested structure of a figure in a deterministic order.

This method takes the place of the Figure.get_ordered method in plotly.py < 3.0.0.

See #1158 for some discussion

* Added to_ordered_dict

* Use setitem rather than setattr syntax in graph object constructors

This has less indirection, and seems to work around a strange
nose testing bug with frames.