Allow forcing a fixed axis scale ratio to maintain quiver arrow lengths (#1197)
A new scaleratio parameter is added to the create_quiver function to allow forcing a fix scale ratio to the axis of the plot through the scaleratio option in the Layout. A distortion is introduced to the arrows length, angle and shape by having a scale ratio different than 1. In this pull request, this distortion is reversed by multiplying or dividing by scaleratio at key steps of the arrows construction.