Add 'cdn' and 'directory' include_plotlyjs options in offline.plot
When 'cdn', the resulting html file/div includes a script tag reference
to the plotlyjs cdn.

When 'directory', the resulting html file/div includes a script tag
reference to a plotly.min.js bundle in the same directory as the html
file. If output_type is 'file' then this plotly.min.js bundle is
created in the output directory if it doesn't already exist.