offline.plot/iplot improvements (#1234)
* Support auto-resize in offline.plot with output_type='div' and include_plotlyjs=False
* Support auto-resize for classic notebook iplot
* Add 'cdn' and 'directory' include_plotlyjs options in offline.plot.  When 'cdn', the resulting html file/div includes a script tag reference
to the plotlyjs cdn. When 'directory', the resulting html file/div includes a script tag
reference to a plotly.min.js bundle in the same directory as the html
file. If output_type is 'file' then this plotly.min.js bundle is
created in the output directory if it doesn't already exist.
* Add option to specify include_plotlyjs as a path/url to a *.js file.  This makes it possible to specify an alternative CDN or offline
location for plotly.js