Update config options and disable show_link by default (#1377)
* Pass through all user-specified config options to Plotly.js and
raise a warning for options that are not known by plotly.py to be valid

Previously we would silently drop unrecognized options.  This change
was done so that if plotly.py falls behind plotly.js in the future,
the user still has a way to specify all config options supported by
plotly.js

* Add missing config options for 1.43.0

* Change show_link default to False.

Along with the change to set `showSendToCloud` to False by default in
plotly.js 1.43.0, now the default behavior of plotly.py/plotly.js is
that there are no options available by default that will send data
off of the local network.