Support calling orca through xvfb on Linux (#1523)
Add orca.config.use_xvfb property to control whether xvfb-run should be used to run orca.  May by True, False, or 'auto'.  The default will be False until v4 (or when the orca_defaults future flag is enabled).  In v4 the default will change to 'auto', in which case xvfb-run is used automatically if plotly.py is running on a Linux machine without a display server with xvfb-run available.