Bad path pointer
When a user tries to write to a bad path, a string describing where in
the path the error occurred is appended to the error message.

Not all the tests are passing because we also want to raise an
appropriate error related to the last valid object in the path. For
example if walking down the path we get to a Line object, but then fail
when trying to look up something in it, we want to
list its valid keys.