Now subscripting errors triggered on types throwing TypeError
Before the subscripting error was only thrown for NoneType objects, but
because stuff like `"hello"["hi"]` and `1["hey"]` also are invalid, and
these also throw TypeError, then these throw the subscripting error
message too. HOWEVER, this error is casted to ValueError on property
assignment and PlotlyKeyError on property read, to keep consistency
among the exception types.