Prevent lazy import when type checking
This checks whether type checking is being done, and if so, disables the
lazy import as this is generally not supported by type checkers.  This
should not have any impact on the runtime performance as the lazy import
should still be used at runtime.

This is achieved by inspecting the
[`TYPE_CHECKING`](https://docs.python.org/3/library/typing.html#typing.TYPE_CHECKING)
variable.

This should resolve issue #1682 and provide a temporary workaround for
issue #1103.  In the latter case, the addition of stubs would probably
still be an improvement as I believe this would be faster.

Signed-off-by: JP-Ellis <josh@jpellis.me>