Update bubble-maps.md
In the python code that sets the variable "limits", the range numbers are resulting in missed values. This is because the first number in the list pair is inclusive while the second number is exclusive. For example, the limit pair of (0,2) only iterates through two loops (0 & 1) instead of three. The net result is that the map doesn't include the "2" city (Chicago).