fixing a bug that causes the hovertext to show %{text} rather than the continent
There is a bug here that causes the hovertext to show one line as %{text} rather than showing the continent name.  I have fixed it the way the comments imply that the example intended.  either we should add a text parameter as I've done here, or the example could remove the %{text} from the hovertext and from the comments.