If IPython is merely installed, don't assume that it's actually used
In my virtual environment, there happens to be installed IPython package.
It wasn't installed by myself, it's merely a transitive dependency of
some another package. It's neither intended to be used, nor is actually
used by my project. However, mere presense of the IPython package, due
to the over-simplistic decision code in this module (io._renderers)
leads to drastic behavior change in Plotly.py: calling
Figure._repr_html_() causes lots of JS garbage to be printed on terminal
(by common sense, this should be a side-effects free function, only
returning a value). On a higher level, this breaks natural Plotly.py
integration in frameworks like Shiny.py (because again, instead of
HTML representation being returned for a framework to process, parts
of that representation are dumped on terminal).

The real underlying issue is that Plotly.py should not force-import
any "optional" packages behind users' back. However, this is a deeper
issue requiring consideration and will be reported separately. This
patch is a quick solution for the problem describe above.

From interactive session:

>>> import IPython
>>> IPython.get_ipython().__class__.__name__
'NoneType'

Signed-off-by: Paul Sokolovsky <pfalcon@users.sourceforge.net>